refactor(rules): extract TLS-rule helpers from android_security.go
Move the helper block at lines 1366-1703 (~338 LOC) backing
OkHTTPDisableSslValidationRule, DisableCertificatePinningRule,
AllowAllHostnameVerifierRule, and InsecureTrustManagerRule into
android_security_tls_helpers.go. Rule structs and check methods
stay in android_security.go and call the helpers as before — same Go
package, no import change.

The generic text utilities matchingParenIndex / firstBraceBody /
stripLineAndBlockComments move with the TLS block; their only other
consumer is android_security_helpers_test.go (same package).

Main file shrinks 1814 → 1474 LOC.
